requirements_inference.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
numpy==1.19.5
packaging==20.8
pandas==1.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
tokenizers==0.9.4
torch==1.7.1
tqdm==4.56.0
transformers==4.2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0
